+++ cloudauth-20190307/ChangeLog.md
@@ -1,3 +1,11 @@
+2024-09-25 Version: 2.9.1
+- Update API CredentialVerify: add param MerchantDetail.
+- Update API CredentialVerify: add param ProductCode.
+- Update API CredentialVerify: add param Prompt.
+- Update API CredentialVerify: add param PromptModel.
+- Update API CredentialVerify: update response param.
+
+
 2024-09-13 Version: 2.9.0
 - Support API InsertWhiteListSetting.
 - Support API PageQueryWhiteListSetting.